refactor: split model CLI commands
